script udpated to handle projects without certificates

+++ migrations/migrationForReportIssue/migrationForReportIssueOfProjects.js
@@ -500,7 +500,7 @@
 
     fs.writeFileSync(masterFilePath, JSON.stringify(masterJsonData, null, 2), "utf8");
 
-    //-----------------------------------------------update the project
+    //-----------------------------------------------update the project-----------------------------------------------
     let projectsToBeUpdated = summary;    
     const programsToBeDeleted = new Set();
     const solutionsToBeDeleted = new Set();
@@ -535,7 +535,7 @@
 
     for(const userEntry of users){
       const {userId, privateProjectIds} = userEntry;
-      // if(["680894a73d8d030008cd0388", "685bd9ac3d8d030008fda1bb"].includes(componentId)) console.log(privateProjectIds)
+
       if(privateProjectIds.length == 0) continue;
       /* Iterate each bug project */
       const projectObjectIds = privateProjectIds.map(id => new ObjectId(id));
@@ -543,16 +543,9 @@
         { _id: { $in: projectObjectIds } },
         { programId: 1, solutionId: 1, status: 1, certificate: 1, tasks : 1, attachments : 1 }
       ).toArray();
-      projects.forEach(project => {
-        if (project.programId) {
-          programsToBeDeleted.add(project.programId.toString());
-        }
-  
-        if (project.solutionId) {
-          solutionsToBeDeleted.add(project.solutionId.toString());
-        }
-  
-      });
+
+      if(projects.length == 0) continue;
+
       const publicProject = await db.collection("projects").findOne(
         {
           solutionId: new ObjectId(componentId),
@@ -563,15 +556,45 @@
           projection: {
             _id: 1,
             tasks: 1,
-            attachments: 1
+            attachments: 1,
+            certificate: 1
           }
         }
       );
+
+      if (!publicProject) {
+        console.log(
+          `No public project found for user ${userId} and component ${componentId}`
+        );
+        continue;
+      }
+
+      // Skip if public project has a certificate & none of the corresponding private-projects has certificate
+      if (
+        publicProject &&
+        publicProject.certificate &&
+        !projects.some(project => project && project.certificate && project.certificate.criteria)
+      ) {
+        continue;
+      }
+
+      projects.forEach(project => {
+        if (project.programId) {
+          programsToBeDeleted.add(project.programId.toString());
+        }
+  
+        if (project.solutionId) {
+          solutionsToBeDeleted.add(project.solutionId.toString());
+        }
+  
+      });
+
+
       // store the private project for which a public project was deleted
       publicToPrivateProjectMap[publicProject._id.toString()] = "";
 
       // deleting public project before converting private project to public
-      if(doUpdate){
+      if((projects.length > 0) && doUpdate){
         await db.collection("projects").deleteOne(
           {_id : publicProject._id}
         )
